+++ src/main/web_src/src/pages/gruppe/PersonListe/modal/KommentarModal.tsx
@@ -1,0 +1,19 @@
+import DollyModal from '~/components/ui/modal/DollyModal'
+import React from 'react'
+import { ErrorBoundary } from '~/components/ui/appError/ErrorBoundary'
+import BeskrivelseConnector from '~/components/beskrivelse/BeskrivelseConnector'
+
+export const KommentarModal = ({ ident, closeModal }) => {
+	console.log(ident) // TODO: slett meg!
+	return (
+		<ErrorBoundary>
+			<DollyModal isOpen closeModal={closeModal} width="50%" overflow="auto">
+				<div className="modal">
+					<h1>Endre kommentar</h1>
+					<br />
+					<BeskrivelseConnector ident={ident} iLaastGruppe={false} />
+				</div>
+			</DollyModal>
+		</ErrorBoundary>
+	)
+}
